frontend/flutter/lib/fastapi_server/app/requirements.txt
fastapi
uvicorn
python-dotenv
groq
gtts
SpeechRecognition